requirements.txt
sanic==22.6.2
sqlalchemy==1.4.41
asyncpg==0.26.0
python_dotenv==0.21.0
PyJWT==2.5.0
bcrypt==4.0.0